Refactor event parser
