+++ tb_plugin/torch_tb_profiler/profiler/event_parser.py
@@ -4,6 +4,7 @@
 import sys
 from collections import defaultdict
 from enum import IntEnum
+from typing import Dict, List
 
 from .. import utils
 from .communication import generate_communication_nodes
@@ -17,18 +18,6 @@ logger = utils.get_logger()
 CommunicationOpNameSet = ['nccl:broadcast', 'nccl:reduce', 'nccl:all_reduce', 'nccl:all_gather', 'nccl:reduce_scatter']
 ProfileRole = IntEnum('ProfileRole', ['Kernel', 'Memcpy', 'Memset', 'Communication', 'Runtime', 'DataLoader', 'CpuOp', 'Other', 'Total'], start=0)
 
-
-class NodeContext:
-    def __init__(self, tid2list, tid2zero_rt_list, corrid_to_device):
-        self.tid2list = tid2list
-        self.tid2zero_rt_list = tid2zero_rt_list
-        self.corrid_to_device = corrid_to_device
-
-class StepContext:
-    def __init__(self, prev_step_end_time, steps_device, steps_matched_device_nodes):
-        self.prev_step_end_time = prev_step_end_time
-        self.steps_device = steps_device
-        self.steps_matched_device_nodes = steps_matched_device_nodes
 
 class NodeParserMixin:
     def __init__(self, *args, **kwargs):
@@ -80,7 +69,7 @@ class NodeParserMixin:
                 logger.warning("{} Runtime with external id {} don't correlate to any operator!".format(
                     len(externalid_to_runtime[ext_id]), ext_id))
 
-        return NodeContext(tid2list, tid2zero_rt_list, corrid_to_device)
+        return tid2list, tid2zero_rt_list, corrid_to_device
 
     def _update_communication_node(self, event):
         '''Update the communication node by using the TraceEvent instance'''
@@ -93,72 +82,6 @@ class NodeParserMixin:
             comm_node.total_time += dur
 
         return comm_node is not None
-
-    def find_device_steps(self, steps):
-        '''return steps associated with device nodes. 
-        '''
-        runtime_node_list = sorted(self.runtime_node_list, key=lambda x: x.start_time)
-
-        # Use similar code with two-way merge to get all runtimes inside each host-side step span,
-        # then record each step's min kernel start time and max kernel end time:
-        steps_device = [(sys.maxsize, -sys.maxsize - 1)] * len(steps)
-        # where the steps associated with devcie node, if yes, the related array item is larger than 0.
-        steps_matched_device_nodes = [0] * len(steps)
-
-        i_step = 0
-        i_runtime = 0
-        step_device_min_ts = sys.maxsize
-        step_device_max_ts = -sys.maxsize - 1
-        matched_device_nodes = set()
-
-        while i_step < len(steps) and i_runtime < len(runtime_node_list):
-            step_host_start_time = steps[i_step][0]
-            step_host_end_time = steps[i_step][1]
-            if runtime_node_list[i_runtime].start_time < step_host_start_time:
-                # This runtime is ahead of or intersects with this step span. Skip this runtime.
-                i_runtime += 1
-            elif runtime_node_list[i_runtime].end_time <= step_host_end_time:
-                # and runtime_node_list[i_runtime].start_time >= step_host_start_time
-                # This runtime is inside this step span. Scan its device_nodes.
-                rt = runtime_node_list[i_runtime]
-                if rt.device_nodes is not None:
-                    for device_node in rt.device_nodes:
-                        step_device_min_ts = min(device_node.start_time, step_device_min_ts)
-                        step_device_max_ts = max(device_node.end_time, step_device_max_ts)
-                        matched_device_nodes.add(device_node)
-                        steps_matched_device_nodes[i_step] += 1
-                i_runtime += 1
-            elif runtime_node_list[i_runtime].start_time < step_host_end_time:
-                # and runtime_node_list[i_runtime].end_time > step_host_end_time
-                # This runtime intersects with this step span. Skip this runtime.
-                i_runtime += 1
-            else:
-                # runtime_node_list[i_runtime].start_time >= step_host_end_time
-                # This runtime starts after this step's end. Record and move forward this step.
-                steps_device[i_step] = (step_device_min_ts, step_device_max_ts)
-                i_step += 1
-                step_device_min_ts = sys.maxsize
-                step_device_max_ts = -sys.maxsize - 1
-
-        while i_step < len(steps):
-            # This step doesn't launch any device side event, just assign it as empty.
-            steps_device[i_step] = (step_device_min_ts, step_device_max_ts)
-            step_device_min_ts = sys.maxsize
-            step_device_max_ts = -sys.maxsize - 1
-            i_step += 1
-
-        # If there are matched device, find the first step end time before steps_device[0][0]
-        prev_step_end_time = None
-        if len(matched_device_nodes) > 0:
-            prev_step_end_time = steps[0][0]
-            if steps_device[0][0] != sys.maxsize:  # When step 0 has device event.
-                for device_node in self.device_node_list:
-                    if device_node not in matched_device_nodes:
-                        # Now this device_node is not launched inside any step span.
-                        if device_node.end_time < steps_device[0][0]:
-                            prev_step_end_time = max(prev_step_end_time, device_node.end_time)
-
-        return StepContext(prev_step_end_time, steps_device, steps_matched_device_nodes)
 
     def _parse_node(self, event, corrid_to_device, corrid_to_runtime, externalid_to_runtime, tid2list, tid2zero_rt_list):
         corrid = event.correlation_id
@@ -211,6 +134,91 @@ class NodeParserMixin:
             tid2list[int(tid)].append(op_node)
 
 
+class OpTreeBuilder:
+    def __init__(self):
+        pass
+
+    def build_tree(self, tid2list, tid2zero_rt_list, corrid_to_device):
+        tid2tree = {}
+
+        staled_device_nodes = []
+        for _, device_nodes in corrid_to_device.items():
+             staled_device_nodes.extend([n for n in device_nodes if n.type == EventTypes.KERNEL])
+
+        for tid, op_list in tid2list.items():
+            zero_rt_list = tid2zero_rt_list[tid] if tid in tid2zero_rt_list else []
+            # Note that when 2 start_time are equal, the one with bigger end_time should be ahead of the other.
+            op_list.sort(key=lambda x: (x.start_time, -x.end_time))
+            main_tid = any([op.name.startswith("ProfilerStep#") for op in op_list])
+            if main_tid:
+                # only append the staled device nodes into main thread
+                root_node = self._build_tree(op_list, zero_rt_list, tid, staled_device_nodes)
+            else:
+                root_node = self._build_tree(op_list, zero_rt_list, tid, [])
+            tid2tree[int(tid)] = root_node
+
+        return tid2tree
+
+    def _build_tree(self, host_node_list, zero_rt_list, tid, device_nodes):
+        '''host_node_list: list of OperatorNode and ProfilerStepNode.
+        zero_rt_list: list of RuntimeNode with external_id=0.'''
+
+        def build_tree_relationship(host_node_list, zero_rt_list, device_nodes):
+            dummpy_rt = []
+            if device_nodes:
+                dummpy_rt.append(RuntimeNode("dummy", 0, 0, EventTypes.RUNTIME, 0, None, 0, device_nodes))
+                dummpy_rt[0].fill_stats()
+            node_stack = []
+            root_node = OperatorNode(
+                name="CallTreeRoot",
+                start_time=-sys.maxsize - 1,
+                end_time=sys.maxsize,
+                type=EventTypes.PYTHON,
+                tid=tid,
+                runtimes=zero_rt_list + dummpy_rt) # Give the list of RuntimeNode with external_id=0 to root node.
+            node_stack.append(root_node)
+            for node in host_node_list:
+                while True:  # break loop when the node is inserted.
+                    tail_node = node_stack[-1]
+                    if node.start_time < tail_node.end_time:
+                        if node.end_time <= tail_node.end_time:
+                            tail_node.children.append(node)
+                            # node.parent_node = weakref.ref(tail_node)
+                            node_stack.append(node)
+                        else:
+                            logger.error("Error in input data: ranges on the same thread should not intersect!"
+                                         "Father:({},{},{}) Child:({},{},{})".format(
+                                tail_node.name, tail_node.start_time, tail_node.end_time,
+                                node.name, node.start_time, node.end_time
+                            ))
+                        break
+                    else:
+                        node_stack.pop()
+            return root_node
+
+        # Merge the consecutive calls to same function into one.
+        # Just follow the same pattern in torch/autograd/profiler.py,
+        # EventList._remove_dup_nodes
+        # TODO: Replace recursive by for loop, in case of too deep callstack.
+        def remove_dup_nodes(node):
+            if node.type == EventTypes.RUNTIME:
+                return
+            if len(node.children) == 1:
+                child = node.children[0]
+                if node.name == child.name and node.type == EventTypes.OPERATOR and child.type == EventTypes.OPERATOR:
+                    node.children = child.children
+                    node.runtimes = child.runtimes  # Keep consistent with autograd profiler.
+                    remove_dup_nodes(node)  # This node may have to merge with child's child.
+            for child in node.children:
+                remove_dup_nodes(child)
+
+        root_node = build_tree_relationship(host_node_list, zero_rt_list, device_nodes)
+        remove_dup_nodes(root_node)
+        root_node.replace_time_by_children()
+        root_node.fill_stats()
+        return root_node
+
+
 class StepParser:
     def __init__(self):
         # we could not use [[]] * len here since they all point to same memory
@@ -249,6 +257,9 @@ class StepParser:
 
         for i in range(len(self.role_ranges)):
             self.role_ranges[i] = merge_ranges(self.role_ranges[i])
+
+    def update_device_steps(self, runtime_node_list):
+        self._update_steps_duration(*self._find_device_steps(runtime_node_list))
 
     @property
     def has_runtime(self):
@@ -298,13 +309,77 @@ class StepParser:
         self.global_min_ts = min(self.global_min_ts, ts)
         self.global_max_ts = max(self.global_max_ts, ts + dur)
 
-    def update_steps_duration(self, context):
+
+    def _find_device_steps(self, runtime_node_list):
+        '''return steps associated with device nodes. 
+        '''
+        runtime_node_list = sorted(self.runtime_node_list, key=lambda x: x.start_time)
+
+        # Use similar code with two-way merge to get all runtimes inside each host-side step span,
+        # then record each step's min kernel start time and max kernel end time:
+        steps_device = [(sys.maxsize, -sys.maxsize - 1)] * len(self.steps)
+        # where the steps associated with devcie node, if yes, the related array item is larger than 0.
+        steps_matched_device_nodes = [0] * len(self.steps)
+
+        i_step = 0
+        i_runtime = 0
+        step_device_min_ts = sys.maxsize
+        step_device_max_ts = -sys.maxsize - 1
+        matched_device_nodes = set()
+
+        while i_step < len(self.steps) and i_runtime < len(runtime_node_list):
+            step_host_start_time = self.steps[i_step][0]
+            step_host_end_time = self.steps[i_step][1]
+            if runtime_node_list[i_runtime].start_time < step_host_start_time:
+                # This runtime is ahead of or intersects with this step span. Skip this runtime.
+                i_runtime += 1
+            elif runtime_node_list[i_runtime].end_time <= step_host_end_time:
+                # and runtime_node_list[i_runtime].start_time >= step_host_start_time
+                # This runtime is inside this step span. Scan its device_nodes.
+                rt = runtime_node_list[i_runtime]
+                if rt.device_nodes is not None:
+                    for device_node in rt.device_nodes:
+                        step_device_min_ts = min(device_node.start_time, step_device_min_ts)
+                        step_device_max_ts = max(device_node.end_time, step_device_max_ts)
+                        matched_device_nodes.add(device_node)
+                        steps_matched_device_nodes[i_step] += 1
+                i_runtime += 1
+            elif runtime_node_list[i_runtime].start_time < step_host_end_time:
+                # and runtime_node_list[i_runtime].end_time > step_host_end_time
+                # This runtime intersects with this step span. Skip this runtime.
+                i_runtime += 1
+            else:
+                # runtime_node_list[i_runtime].start_time >= step_host_end_time
+                # This runtime starts after this step's end. Record and move forward this step.
+                steps_device[i_step] = (step_device_min_ts, step_device_max_ts)
+                i_step += 1
+                step_device_min_ts = sys.maxsize
+                step_device_max_ts = -sys.maxsize - 1
+
+        while i_step < len(self.steps):
+            # This step doesn't launch any device side event, just assign it as empty.
+            steps_device[i_step] = (step_device_min_ts, step_device_max_ts)
+            step_device_min_ts = sys.maxsize
+            step_device_max_ts = -sys.maxsize - 1
+            i_step += 1
+
+        # If there are matched device, find the first step end time before steps_device[0][0]
+        prev_step_end_time = None
+        if len(matched_device_nodes) > 0:
+            prev_step_end_time = self.steps[0][0]
+            if steps_device[0][0] != sys.maxsize:  # When step 0 has device event.
+                for device_node in self.device_node_list:
+                    if device_node not in matched_device_nodes:
+                        # Now this device_node is not launched inside any step span.
+                        if device_node.end_time < steps_device[0][0]:
+                            prev_step_end_time = max(prev_step_end_time, device_node.end_time)
+
+        return prev_step_end_time, steps_device, steps_matched_device_nodes
+
+
+    def _update_steps_duration(self, prev_step_end_time, steps_device, steps_matched_device_nodes):
         '''Update self.steps considering device side events launched by each host side step.
         Update self.steps_names if some tail steps are removed.'''
-
-        prev_step_end_time = context.prev_step_end_time
-        steps_device = context.steps_device
-        steps_matched_device_nodes = context.steps_matched_device_nodes
 
         # Change step time to device side on the condition that any step have device time.
         is_use_gpu = prev_step_end_time is not None
@@ -345,18 +420,18 @@ class StepParser:
                     self.steps = self.steps[:keep_steps]
                     self.steps_names = self.steps_names[:keep_steps]
 
-class EventParser(NodeParserMixin, StepParser):
+class EventParser(NodeParserMixin, StepParser, OpTreeBuilder):
     def __init__(self):
         super().__init__()
 
-    def parse(self, events):
-        node_context = self.parse_nodes(events)
-        self.parse_steps(events, self.communication_data)
+    def parse(self, events) ->  Dict[int, List[OperatorNode]]:
+        tid2tree = self.build_tree(*self.parse_nodes(events))
 
+        # Process steps
+        self.parse_steps(events, self.communication_data)
         # Move the interleaved logic out of each NodeParser and StepParser
-        steps_context = self.find_device_steps(self.steps)
-        self.update_steps_duration(steps_context)
-        return node_context
+        self.update_device_steps(self.runtime_node_list)
+        return tid2tree
 
     def generate_communication_nodes(self):
         return generate_communication_nodes(self.communication_data, self.steps, self.steps_names)
